Renamed a few packages to internal in the sample bundles. See openhab/static-code-analysis#100
