+++ bundles/automation/org.eclipse.smarthome.automation.sample.extension.java/src/main/java/org/eclipse/smarthome/automation/sample/extension/java/internal/template/WelcomeHomeTemplateProvider.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.automation.sample.extension.java.template;
+package org.eclipse.smarthome.automation.sample.extension.java.internal.template;
 
 import java.util.Collection;
 import java.util.Collections;
