+++ bundles/automation/org.eclipse.smarthome.automation.sample.rest.api/src/main/java/org/eclipse/smarthome/automation/sample/rest/api/internal/Activator.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.automation.sample.rest.api;
+package org.eclipse.smarthome.automation.sample.rest.api.internal;
 
 import java.io.IOException;
 import java.net.URL;
